+++ src/plugins/blocksmith/components/vimeo-video-block-modal/index.jsx
@@ -83,10 +83,10 @@ class VimeoVideoBlockModal extends React.Component {
     this.handleChangeDate = this.handleChangeDate.bind(this);
     this.handleChangeTextArea = this.handleChangeTextArea.bind(this);
     this.handleChangeTime = this.handleChangeTime.bind(this);
+    this.handleClearImage = this.handleClearImage.bind(this);
     this.handleEditBlock = this.handleEditBlock.bind(this);
     this.handleSelectImage = this.handleSelectImage.bind(this);
-    this.handleToggleAssetPickerModal = this.handleToggleAssetPickerModal.bind(this);
-    this.handleClearImage = this.handleClearImage.bind(this);
+    this.handleToggleImageAssetPickerModal = this.handleToggleImageAssetPickerModal.bind(this);
   }
 
   setBlock() {
@@ -240,12 +240,12 @@ class VimeoVideoBlockModal extends React.Component {
     });
   }
 
-  handleToggleAssetPickerModal() {
-    this.setState(({ isAssetPickerModalOpen }) => ({
-      isAssetPickerModalOpen: !isAssetPickerModalOpen,
+  handleToggleImageAssetPickerModal() {
+    this.setState(({ isImageAssetPickerModalOpen }) => ({
+      isImageAssetPickerModalOpen: !isImageAssetPickerModalOpen,
     }), () => {
-      const { isAssetPickerModalOpen } = this.state;
-      if (!isAssetPickerModalOpen) {
+      const { isImageAssetPickerModalOpen } = this.state;
+      if (!isImageAssetPickerModalOpen) {
         this.refocusModal();
       }
     });
@@ -276,7 +276,7 @@ class VimeoVideoBlockModal extends React.Component {
       errorMsg,
       hasUpdatedDate,
       id,
-      isAssetPickerModalOpen,
+      isImageAssetPickerModalOpen,
       isValid,
       selectedImageNode,
       touched,
@@ -290,7 +290,7 @@ class VimeoVideoBlockModal extends React.Component {
     const { isFreshBlock, isOpen, node, toggle, blockKey } = this.props;
 
     return (
-      <Modal isOpen={isOpen} toggle={toggle} autoFocus={false} keyboard={!isAssetPickerModalOpen}>
+      <Modal isOpen={isOpen} toggle={toggle} autoFocus={false} keyboard={!isImageAssetPickerModalOpen}>
         <ModalHeader toggle={toggle}>{`${isFreshBlock ? 'Add' : 'Update'} Vimeo Video Block`}</ModalHeader>
         <ModalBody>
           <FormGroup>
@@ -313,15 +313,15 @@ class VimeoVideoBlockModal extends React.Component {
           }
           <FormGroup>
             <ImageAssetPicker
-              multiAssetErrorMessage="Invalid Action: Trying to assign multiple Vimeo Block Poster images."
-              isImageSelected={!!selectedImageNode}
-              isModalOpen={isAssetPickerModalOpen}
               isDisabled={!isValid || !id}
+              isImageSelected={!!selectedImageNode}
+              isModalOpen={isImageAssetPickerModalOpen}
               label="Select A Vimeo Block Poster Image"
+              multiAssetErrorMessage="Invalid Action: Trying to assign multiple Vimeo Block Poster images."
               node={node}
               onClearImage={this.handleClearImage}
               onSelectImage={this.handleSelectImage}
-              onToggleAssetPickerModal={this.handleToggleAssetPickerModal}
+              onToggleImageAssetPickerModal={this.handleToggleImageAssetPickerModal}
             />
           </FormGroup>
           <FormGroup>